Added +copy and link previews

+++ bot/cogs/admin.py
@@ -24,7 +24,7 @@ class Admin(commands.Cog):
         embed.add_field(name="# Text channels", value=n_text, inline=False)
         await ctx.send(embed=embed)
 
-    @commands.command()
+    @commands.command(aliases=["pong"])
     async def ping(self, ctx):
         await ctx.send("Ponguu!")
 
@@ -108,9 +108,11 @@ class Admin(commands.Cog):
         await asyncio.sleep(3)
         await reply.delete()
 
-    @commands.command()
+    @commands.command(aliases=["copy"])
     @commands.has_any_role("Moderator", "Guides of the Void")
     async def move(self, ctx, msg_id, to_channel, from_channel=None):
+
+        cmd = re.split(" ", ctx.message.content)[0][1:]
 
         await ctx.message.delete()
 
@@ -125,29 +127,54 @@ class Admin(commands.Cog):
             raise commands.errors.BadArgument(
                 "Please enter valid channel names, no leading #"
             )
-        else:
-            try:
-                msg = await from_channel.fetch_message(int(msg_id))
-            except Exception:
-                raise commands.errors.BadArgument("No message found with that ID")
+        try:
+            msg = await from_channel.fetch_message(int(msg_id))
+        except Exception:
+            raise commands.errors.BadArgument("No message found with that ID")
 
-            embed = discord.Embed(color=PALE_BLUE)
-            embed.title = "Moved message"
-            embed.set_thumbnail(url=msg.author.avatar_url)
-            embed.description = msg.content
-            embed.add_field(
-                name="Original poster", value=msg.author.mention, inline=True
-            )
-            embed.add_field(name="Channel", value=msg.channel.mention, inline=True)
-            timestamp = msg.created_at + datetime.timedelta(hours=2)
-            timestamp = timestamp.strftime("%b %d, %H:%M")
-            embed.add_field(name="Sent at", value=timestamp, inline=True)
-            if msg.attachments:
-                embed.set_image(url=msg.attachments[0].url)
-            elif re.search(r"^https:.*\.(jpe?g|png|gif)$", msg.content):
-                embed.set_image(url=msg.content)
-            await to_channel.send(embed=embed)
+        embed = message_embed(msg, cmd, ctx.author)
+
+        await to_channel.send(embed=embed)
+        if cmd == "move":
             await msg.delete()
+
+
+def message_embed(msg, cmd, trigger_author) -> discord.Embed:
+
+    embed_titles = {
+        "copy": "Copied message",
+        "move": "Moved message",
+        "link": "Message preview",
+    }
+
+    trigger_field_titles = {
+        "move": "Moved by",
+        "copy": "Copied by",
+        "link": "Linked by",
+    }
+
+    embed = discord.Embed(color=PALE_BLUE)
+    embed.title = embed_titles[cmd]
+    embed.set_thumbnail(url=msg.author.avatar_url)
+    embed.description = msg.content
+    embed.add_field(name="Channel", value=msg.channel.mention)
+    timestamp = msg.created_at + datetime.timedelta(hours=2)
+    timestamp = timestamp.strftime("%b %d, %H:%M")
+    embed.add_field(name="Sent at", value=timestamp)
+    embed.add_field(name="\u200b", value="\u200b")
+    embed.add_field(name="Original poster", value=msg.author.mention)
+
+    embed.add_field(name=trigger_field_titles[cmd], value=trigger_author.mention)
+
+    if cmd == "copy" or cmd == "link":
+        embed.add_field(name="Link to message", value=f"[Click here]({msg.jump_url})")
+
+    if msg.attachments:
+        embed.set_image(url=msg.attachments[0].url)
+    elif re.search(r"^https:.*\.(jpe?g|png|gif)$", msg.content):
+        embed.set_image(url=msg.content)
+
+    return embed
 
 
 def setup(bot):
